rename: promptolution.experiment_grid -> promptolution.experiment
"grid" undersold the module: it runs single experiments too, not just
parameter sweeps. Renames the package, its test directory and docs page,
and updates every path/import string (pyproject include, mkdocs nav,
docs/index.md, mkdocstrings target, test imports and patch targets,
release notes).

Also: drop the demo task's old "dummy" name (config.yaml default and test
overrides -> demo), drop the qos: mcml line from the slurm launcher config
(leftover from the original cluster), and add a header comment to
task/csv.yaml explaining what it is for and its `???` required-value
markers.
