+++ docs/api/experiment.md
@@ -1,0 +1,5 @@
+# Experiment
+
+::: promptolution.experiment.launch_grid
+    options:
+      show_root_heading: true
