--- docs/api/experiment_grid.md
@@ -1,5 +1,0 @@
-# Experiment Grid
-
-::: promptolution.experiment_grid.launch_grid
-    options:
-      show_root_heading: true
