+++ docs/release-notes/v2.3.0.md
@@ -6,8 +6,8 @@ Promptolution v2.3.0 removes ExperimentConfig and its string dispatch get\_\* fa
 
 #### Added features:
 
-- `promptolution.experiment_grid`: config-driven single runs and whole experiment grids via Hydra
-  `instantiate`: `python -m promptolution.experiment_grid`, `-m` for grids, one SLURM array job via `hydra/launcher=slurm`. Every component group is a `_target_` + its params in `conf/`.
+- `promptolution.experiment`: config-driven single runs and whole experiment grids via Hydra
+  `instantiate`: `python -m promptolution.experiment`, `-m` for grids, one SLURM array job via `hydra/launcher=slurm`. Every component group is a `_target_` + its params in `conf/`.
 - Restart for experiments: the experiment `name` keys the output folder; rerunning the same name
   resumes into the same folder and skips cells that already finished.
 - Per-run output: `step_results.parquet`, `prompt_scores.parquet`, `runinfo.json`, and a
@@ -20,7 +20,7 @@ Promptolution v2.3.0 removes ExperimentConfig and its string dispatch get\_\* fa
 
 - `promptolution.helpers` (`run_experiment`, `run_optimization`, `run_evaluation`, `get_llm`,
   `get_task`, `get_optimizer`, `get_predictor`, `get_exemplar_selector`) and `ExperimentConfig` are
-  removed. Build components directly (see the README quickstart) or use `experiment_grid`.
+  removed. Build components directly (see the README quickstart) or use `experiment`.
 - Every component constructor drops its `config` parameter.
 - `hydra-core` is now a main dependency; the `[experiments]` extra keeps the SLURM launcher plugin
   and HuggingFace `datasets`.
