+++ mkdocs.yml
@@ -73,7 +73,7 @@ nav:
     - Optimizers: api/optimizers.md
     - Predictors: api/predictors.md
     - Tasks: api/tasks.md
-    - Experiment Grid: api/experiment_grid.md
+    - Experiment: api/experiment.md
     - Utils: api/utils.md
     - Exemplar Selectors: api/exemplar_selectors.md
   - Tutorials:
