+++ promptolution/experiment/__init__.py
@@ -5,10 +5,10 @@
 and writes the per-run output contract (results, runinfo, restart marker) — see `README.md`.
 
 Usage:
-    python -m promptolution.experiment_grid name=my_run optimizer=capo task=agnews llm=api  # one run
-    python -m promptolution.experiment_grid -m name=bench optimizer=capo,opro random_seed=42,43  # a grid
+    python -m promptolution.experiment name=my_run optimizer=capo task=agnews llm=api  # one run
+    python -m promptolution.experiment -m name=bench optimizer=capo,opro random_seed=42,43  # a grid
 """
 
-from promptolution.experiment_grid.launch_grid import execute
+from promptolution.experiment.launch_grid import execute
 
 __all__ = ["execute"]
