+++ tests/experiment/test_grid.py
@@ -1,11 +1,11 @@
 """Restart behavior of execute(): name-keyed dir, skip finished cells."""
 
-from promptolution.experiment_grid import execute
+from promptolution.experiment import execute
 
 
 def test_restart_skips_finished(cfg_with_mock_llm, tmp_path):
     """A second run into the same folder skips the finished cell (does not re-execute)."""
-    cfg = cfg_with_mock_llm(["name=t", "task=dummy", "optimizer=evopromptga", "n_steps=2", "test_frac=0.25"])
+    cfg = cfg_with_mock_llm(["name=t", "task=demo", "optimizer=evopromptga", "n_steps=2", "test_frac=0.25"])
     execute(cfg)
     assert (tmp_path / ".finished").exists()
     (tmp_path / "step_results.parquet").unlink()  # remove an output
